fix: eager-materialize workspace drag payload
